Partial Tree Fix - Stream Fix
More than one issue has cropped up.  First is squashed with this commit - selective mode would leave whitespace, if it was there, at the end of lines as it thought it had encountered an encapsulation.

The other problem is more tedious - the whole parse function might be broken.  I am getting nodes in the wrong places - main issue seems to be searching for parent nodes, which results in either ending up in the wrong place or crashing the vm.